Make errors more descriptive than "The API returned an invalid content type".
Where we call `AssertJsonContentOrEmpty` prior to doing a status code check, in the case of the API returning an HTML error page (e.g., for an HTTP 401/403/404/429/503) then the error displayed in C# will be an unhelpful "An error has occured" combined with an error guid and to figure out the cause requires going into the logging.

Moving this JSON check to after the status code check allows for the correct error details to be shown.

Example: HTTP 401 in the API leads to an unhelpful "invalid content type" error as opposed to an error about the API connection being unauthorised.