Failing validation adding chair of trustees causes the "Original chair of trustees" to be visible for a MAT where it should not be. (#475)
The controller method that calls the java api to add this now has the code that modifies the display policy where the group is not secure 16-19.  This makes the field invisible where the group is not type 11.  The code originally had this on the get method for AddEditOrReplace now it is also included on the post.

This was required because the group type is not one of the parameters for the api.  Each time the display policy is reloaded from the api this modification would be required or the values reset to true.  It appears this was done to save adding the extra paramater to the java api.